fix: split monolithic sitemap into per-family route handlers
The single app/sitemap.ts with generateSitemaps() bundled ALL data-loading
imports (courses, transfer, programs, instructors, blog, online) into ONE
serverless function. With 157 MB of course data + 175 MB of transfer data,
this exceeded Vercel's 250 MB unzipped function size cap.

Split into 9 independent route handlers under app/sitemap/{family}.xml/,
each importing only its own data. Add a sitemap index at app/sitemap.xml/
and simplify robots.ts to point at the index.
