+++ app/robots.ts
@@ -1,18 +1,5 @@
 import type { MetadataRoute } from "next";
 
-// Keep in sync with SITEMAP_IDS in app/sitemap.ts.
-const SITEMAP_IDS = [
-  "core",
-  "colleges",
-  "college-subjects",
-  "courses",
-  "state-subjects",
-  "transfer",
-  "instructors",
-  "programs",
-  "blog",
-];
-
 export default function robots(): MetadataRoute.Robots {
   const baseUrl =
     process.env.NEXT_PUBLIC_SITE_URL || "https://communitycollegepath.com";
@@ -22,6 +9,6 @@ export default function robots(): MetadataRoute.Robots {
       allow: "/",
       disallow: ["/api/"],
     },
-    sitemap: SITEMAP_IDS.map((id) => `${baseUrl}/sitemap/${id}.xml`),
+    sitemap: `${baseUrl}/sitemap.xml`,
   };
 }
